Create Of the given 5 numbers, How many are even or odd?
Write a function that accepts 6 input parameters. The first 5 input parameters are of type int.
The sixth input parameter is of type string. If the sixth parameter contains the value "even", 
the function is supposed to return the count of how many of the first five input parameters are
even. If the sixth parameter contains the value "odd", the function is supposed to return the
count of how many of the first five input parameters are odd.
For example,
If the five input parameters are 12, 17, 19, 14, and 115, and the sixth parameter is "odd",
the function must return 3 because there are three odd numbers 17, 19, and 115.

If the five input parameters are 12, 17, 19, 14, and 115, and the sixth parameter is "even", the
the function must return 2 because there are two even numbers 12 and 14.

Note that zero is considered an even number.