Create FACTORIAL of a number
In mathematics, the factorial of a non-negative integer n, denoted by n! is the product of all positive integers less than or equal to n. For example,
5!= 5 X 4 X 3 X 2 X 1 = 120
5!= 4 X 3 X 2 X 1 = 24
9!= 9 X 8 X 7 X 6 X 5 X 4 X 3 X 2 X 1 = 362880
Write a program to find the factorial of a given number.
The given number will be passed to the function as an input parameter of type int.
The function is expected to calculate the factorial of the given number and return it as an int type.
The assumption for this program:
The given input number will always be greater than or equal to 1.
Due to the range supported by int, the input number will range from 1 to 12.